backend/requirements.txt
fastapi==0.110.0
uvicorn==0.28.0
psutil==5.9.8
pymongo==4.6.1
python-socketio==5.11.0
websockets==12.0
python-dotenv==1.0.1
pydantic==2.6.1
motor==3.3.2